requirements.txt
altgraph==0.17.4
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
oauthlib==3.2.2
packaging==24.2
pefile==2023.2.7
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.1.4
pytz==2024.2
pywin32-ctypes==0.2.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
